% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan1d-package.R
\docType{package}
\name{dbscan1d-package}
\alias{dbscan1d-package}
\title{Efficient specialization of DBSCAN clustering algorithm for 1D data}
\description{
A O(nlogn) pessimistic complexity implementation of DBSCAN
(Density Based Clustering of Applications with Noise) optimized for 1D data.
If data is already sorted the algorithm runs in (n) time.
}
