% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbscan1d}
\alias{dbscan1d}
\title{Efficent implementation of DBSCAN alghorithm for 1D data}
\usage{
dbscan1d(data, eps, minPoints, sort = TRUE)
}
\arguments{
\item{data}{numeric vector with points to cluster}

\item{eps}{the epsilon-neighborhood value for clustering}

\item{minPoints}{the number of points for core-point criterion}

\item{sort}{whether to sort the data, pass FALSE if calling for sorted data}
}
\value{
a vector of cluster labels
}
\description{
Perfoms clustering with DBSCAN alghorithm optimized for one dimensional data.
For sorted data (pass FALSE to sort parameter) the complexity is O(n) where n is the length of the data vector.
If sorting is needed the algorithm complexity is O(nlogn).
}
\examples{
dbscan1d(c(9,8,7, 1,2,3, 15,16,17), eps=1, minPoints=2)

}
